Replace hemi call patterns
